data_engineering_2/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
kafka-python==2.0.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
